feat: secure sensitive endpoints

+++ orchestration-api/modules/users/UsersController.js
@@ -1,5 +1,6 @@
 const { Router } = require('express');
 const { getCollection, getDocument, setDocument, updateDocument, deleteDocument } = require('../../clients/FirebaseClient');
+const { getUserRole } = require('../../lib/SecurityService');
 
 const router = Router();
 
@@ -30,15 +31,16 @@ router.get('/', async (req, res, next) => {
 
 router.post('/', async (req, res, next) => {
 	try {
-		const { user, userId } = req.body;
+		const { networkId } = req.user;
+		const user = req.body;
 
-		const [{ networkId }, users] = await Promise.all([
-			getDocument(`users/${userId}`),
-			getCollection('users')
-		]);
+		const requestUserRole = await getUserRole(req.user._id, networkId);
+		if (!['admin', 'owner'].includes(requestUserRole)) throw new Error('Not authorized');
+
+		const users = await getCollection('users');
 
 		const newUser = users.find(u => u.email === user.email);
-		if (!newUser) throw new Error('user not found');
+		if (!newUser) throw new Error('user does not exist in home-orchestrator');
 		user.userId = newUser._id;
 
 		const result = await setDocument(`/networks/${networkId}/users/${user.userId}`, user);
@@ -53,6 +55,13 @@ router.patch('/:id', async (req, res, next) => {
 		const user = req.body;
 		const { networkId } = req.user;
 		const { id } = req.params;
+
+		const requestUserRole = await getUserRole(req.user._id, networkId);
+		if (!['admin', 'owner'].includes(requestUserRole)) throw new Error('Not authorized');
+
+		const { ownerId } = await getDocument(`networks/${networkId}`);
+		if (ownerId === id) throw new Error('Cannot modify network owner');
+
 		const result = await updateDocument(`networks/${networkId}/users/${id}`, user);
 		return res.send(result);
 	} catch (err) {
@@ -64,6 +73,13 @@ router.delete('/:id', async (req, res, next) => {
 	try {
 		const { networkId } = req.user;
 		const { id } = req.params;
+
+		const requestUserRole = await getUserRole(req.user._id, networkId);
+		if (!['admin', 'owner'].includes(requestUserRole)) throw new Error('Not authorized');
+
+		const { ownerId } = await getDocument(`networks/${networkId}`);
+		if (ownerId === id) throw new Error('Cannot delete network owner');
+
 		const result = await deleteDocument(`networks/${networkId}/users/${id}`);
 		return res.send(result);
 	} catch (err) {
